+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.11.5
 
 * Force initialization of all tables in catalog in `build_compute`
+* Do not throw `NotImplementedError` in `run_changelist`
 
 # 0.11.4